setup redux store and injected into index.js
